fix(pr3-review-r2): address review feedback
Round-2 fixes for PR-3 in response to code-reviewer (REQUEST_CHANGES,
2 majors + 4 minors), generative-protein-scientist (3 minors), and
ml-software-pytorch-jax-expert (3 minors):

- Drop `name: plddt` key from plddt_head.yaml. The key was breaking
  the dominant codebase pattern of calling `hydra.utils.instantiate(cfg)`
  directly on a head config: PLDDTHead.__init__ doesn't accept name.
  Tests for the name-style factory path now add the key in-test.
- Update BaseConfidenceHead.forward chain_id docstring: was "defaults to
  zeros (monomer)" but the code only del'd it. Now says "accepted but
  currently ignored; reserved for future ipTM/ipAE/ipLDDT subclasses".
- Drop the unused ConfidenceTrunk.expects_external_cond kwarg/attribute
  and its base.yaml entry (also dropped from trunk tests).
- Drop the redundant bin_centers.to(probs.dtype) no-op cast in
  PLDDTHead.logits_to_expected_value (bin_centers is already fp32).
- Add tests/unit/confidence/test_projections.py covering the identity
  branch and dim-changing branch of SeqProjection / PairProjection,
  plus mask consistency.
- Add directional calibration test for the head EV: saturated logit on
  bin index 30 (center 61.0) -> EV within 0.5 of 61.0.

53/53 tests pass.

+++ src/proteinfoundation/nn/confidence/plddt_head.py
@@ -64,4 +64,4 @@ class PLDDTHead(BaseConfidenceHead):
         """Softmax-weighted bin-center mean. fp32 internally."""
         logits_f = logits.float()
         probs = torch.softmax(logits_f, dim=-1)
-        return (probs * self.bin_centers.to(probs.dtype)).sum(dim=-1)
+        return (probs * self.bin_centers).sum(dim=-1)
